pyproject.toml
[project]
name = "sudar-agent"
version = "1.0.0"
description = "AI-powered Educational Assistant for Multi-grade Classrooms"
requires-python = ">=3.10"

dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.12.0",
    "langchain-docling",
    "langchain-huggingface",
    "langchain-community",
    "langchain-ollama",
    "qdrant-client[fastembed]",
    "langchain-google-genai",
    "langchain-groq",
    "tavily-python",
    "python-dotenv",
    "rich",
    "requests",
    "sentence-transformers",
    "pymongo",
    "minio",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]
